Index Is Now Most Recent Article
Dropped stock FontAwesome css in favor of trim cod sources and included
FontAwesome directly as a web font.

Multiple base structure changes, most notably moving articles list to a
footer item and making the index display the most recent article. Also
moved social links a little in the header.

Changed article titles to h1s.

Re-introduced the features to title the links and social menus, and
added a similar feature for the feed link in the social menu. (TODO:
Improve Conditionals for Feeds in Social Menu)

Moved pagination to includes where it belongs.