Removed some unused hover styles from FontAwesome and added a few more icons.
Several page structure improvements, especially conforming font sizes
to scale based on the attribute set at <html>

Moved the contents of the <article> to a separate include to stay DRY.

Improved category layout to extend base and provide a proper list. Did
the same for tags.

Fixed an issue with the articles list wherein it was being filtered
when a particular topic or tag is selected. Picking from all_articles
isn’t good enough; it can only be used when articles is filtered!

Added a class to comment count links.

Set blogroll links to open in a new window.